requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
huggingface-hub>=0.17.0

